Fix [#66758184] by switching to Query API
The Fog::ServiceInterface code for finding a vAppTemplate
mistakenly found ISOs with similar names.

This commit updates the Core::VappTemplate.get factory to
use the Query API to find the vAppTemplate id explicitly.

Tests are updated accordingly, but remain equivalent to the
previous test. More to be added in suring up commits to follow.