These tests are not being run
Because of the way these tests are written, they only run if the config is there, so they were 'failing' silently. They would have failed anyway because this test is now running in vcloud-core, where there is no `examples` directory.

I am not convinced of the value of these tests in any case. They are there to check our example documentation is up-to-date, but it's not clear that the example documentation is linked to or used. If it is deemed to be useful, these tests should be reimplemented in their respective gems.