Split out query functionality into two classes
The Vcloud::Query class was combining two separate concerns -
handling the command line inputs/formatting the output and
interacting with Fog. This commit splits out those two pieces of
functionality. The Query class handles the command line input and
formats the output and the QueryRunner class does the interactions
with Fog.

Also removed the code for debug output, which wouldn't have worked
anyway due to no require for 'pp' and the fog mock code.