Upgrade Fog version to fix progress bar issue
The upgrade of our supplier to 5.5 exposed an issue with the task progress bar in that it wasn't returned quickly enough so the code fell over. A fix was made in Fog https://github.com/fog/fog/pull/2837 and this commit bumps the required Fog version to take advantage of the fix.