Remove confusing vcloud-query 'record type' column
vcloud-query with no options would return a list of 'queriable
Entity types', along with a comma-separated list of 'record formats'
available for this particular entity.

We only ever use the 'records' format. The other types just add noise.

Simplify the output of the tool to just list entity types
available in 'records' format.