Update available_query_types to be a simple list
We only care about querying the API for 'format=records'. The
other two formats ('references' and 'idrecords') are cluttering our
code.

This commit reduces the output of available_query_types to be a simple
array of entity types that are queriable in records format.

As a result, the output_available_query_types and associated tests are
updated to reflect the new return value.