Ensure EdgeGateway integ tests have correct ENV
Our EdgeGateway integration tests require certain environment variables
to be set. These are documented in the README.md, but if this step is
missed then the tests fail unintuitively.

This commit adds the env checking code from other tests to ensure that
these are set before the tests are run.