Add var_files arg to load_config and failing test
Accept an optional `vars_file` argument to #load_config which is the path of
a YAML file containing some variables which in turn causes the `config_file`
to be treated as a template and string interpolation is performed.

This will allow tools like vcloud-edge_gateway to re-use a single ruleset
for multiple environments where the only thing that changes are IP addresses
stored in separate files. It can also be re-used by some of our unit tests
which currently use spec helpers to do the same job.

Actual implementation to follow in a separate commit.