Use Mustache to template ConfigLoader files
Satisifies the failing test from the previous commit.

This has the advantage over ERB that the template files are completely
logic-less and you can't do anything exciting like loops or querying
external data sources, which might render your deployments unreproducible.

Per the description at:
- http://mustache.github.io/mustache.5.html#DESCRIPTION

I'm not attempting to support the YAML frontmatter method of storing the
variables and template in the same file like `mustache(1)` does because we
already have (and test) YAML anchors for this:

- http://mustache.github.io/mustache.1.html#DESCRIPTION