[#70513870] Standardise testing section in README
I plan to use this same text in other projects which don't currently have a
section about testing. I don't think it's useful to describe all of the
types of tests that we have, because they will change over time and the
documentation will invariably be out of date. We should just document how to
run the default set of tests and how to run things that aren't covered by
the default.