Restrict available scope in ERB preamble
Previously we were giving ERB the binding object of Vm.self, which
meant that all variables in the scope of the Vm object were available,
plus any methods on the Vm object itself.

This is kind of risky, and allows people to do operations in the ERB
that we are not comfortable with.

Instead, create an OpenStruct object from our vars hash supplied, and
vapp_name (since this is used in real-world templates). Then supply
the binding object of this object, which is far more limited.