Move from deprecated get_network call
... to get_network_complete.

The get_network request rewrites the output from the vCloud API, and loses information
in the process. It has been deprecated in Fog since 1.21.

The discussion surrounding this deprecation is in https://github.com/fog/fog/pull/2683

As a result, this commit alters the output of OrgVdcNetwork#vcloud_attributes,
making it more closely reflect the output from the vCloud API, rather than the
re-written output of get_network.

At the moment, the only place known to be using vcloud_attributes are the
integration tests for http://github.com/alphagov/vcloud-net_launcher