Increase the footprint of the postprocessor test
This commit improves the test on the preamble ERB optional
postprocessor script. Previously we used /bin/cat, so in effect were
using a no-op postprocessor. In addition, the re-use of the ERB
from the previous 'interpolation' test meant we were duplicating
test code.

This test instead tests purely the post-processor function, ensuring
that the character count in the source file matches the output of 'wc'
(word count -- also reports character count). This ensures that the
post-processor *must* run for the test to succeed.