Change arguments for VappTemplate#get
Swap the order of the template and catalog arguments so that they match the
`#get_ids_by_name_and_catalog` method, which can't easily be changed because
the order is reflected in the method name, and besides it reads more
naturally this way round - you want to get a specific item (first, can never
be a default) from a broader collection (second, could be a default catalog).

We temporarily decided not to so this because it means a matching version
change in vcloud-launcher. But we'll be releasing feature changes to schema
validation, so we might as well do this now.

Renaming `catalog_item_name` to `vapp_template_name` to match renaming about
to occur in vcloud-launcher. I'm not very attached to the `vapp_template_`
prefix because we're already in the VappTemplate class.