Add Vm#configure_network_interfaces integ tests
Test that we can create single and dual nic systems.

This highlights that Fog is misbehaving wrt the vcloud_attributes
section for network cards -- ensure_list! is missing, so we get a
Hash if there is one NIC, an Array if there are >1.

On testing against a second provider, it was noticed that
the :NetworkConnection section returned by the API via Fog is
not ordered as per the insertion order.

Hence, add a helper (sort_nics_based_on_network_connection_index) to sort the
returned array based on the :NetworkConnectionIndex parameter, which then allows
us to test each returned hash in respect of the order we used to insert them.