Skip Vm#configure_guest_customization_section tests
Vm#configure_guest_customization_section contains and calls a lot of
code that handles the 'preamble script' generation, including
reading this from an external file.

This work is better done in Vcloud::Launch::VmOrchestrator, with
an interpolated preamble passed to core as data.

Leaving this test alone for now until we decide what to do, there
are tests in alphagov/vcloud-launcher for this already, so there is
no major reason to re-implement them here right now.