Add Vapp.instantiate passing integ tests
This confirms the behaviour that we can instantiate vAppTemplates
as vApps with zero, one or two 'bridged networks'.

Theoretically we could test more, but we do not have any user need
for >2 networks as yet.

NB: Adjusted the use of the test_data hash to reduce the need for
instance variables.