Remove vcloud-tools-testing-config after use
CI is failing intermittently because the workspace isn't deleted
before/after each build and the previous clone of the config repo already
exists.

    fatal: destination path 'vcloud-tools-testing-config' already exists and is not an empty directory.
    Build step 'Execute shell' marked build as failure

Prevent this from happening by deleting the repo/dir after we've moved the
one file out of it. We could also amend the `git clean` to `-ffdx`, but I
think it's better to be explicit here.