Simplify edge gateway integration tests
Simplify the edge gateway integration tests such that we only test the
functionality implemented in vCloud Core.

After discussion with @annashipman and @dcarley, we decided that the
existing tests were doing too much by duplicating some of the tests in
vCloud Edge Gateway. Instead, we should only test the functionality
contained in this gem.

Replace the existing test with one test that asserts whether the
EdgeGateway#update_configuration method successfully updates the edge
gateway.

Note that we perform three separate assertions on the configuration
returned by the edge gateway on each of the `:FirewallService`,
`:NatService` and `:LoadBalancerService` hash members. This is necessary
because the edge gateway returns additional information such as
`:GatewayIpsecVpnService` and `:GatewayDhcpService` which we're not
concerned with.