Reset edge gateway configuration before tests
Reset the vShield Edge Gateway's configuration before each test when
necessary to ensure that we are testing against a known configuration.

This adds the `IntegrationHelper::reset_edge_gateway` helper method
which is based on a method of the same name in the vCloud Edge Gateway
gem[1]. I think we can refactor the Edge Gateway gem in future to use
the method I'm adding here to reduce duplication and keep this logic in
vCloud Core.

I've also added a context to the Rspec test concerned to help separate
it from the other tests.

Note that I intentionally don't reset the edge gateway configuration
after the integration tests have run, unlike the vCloud Edge Gateway
gem[2]. There should be no need for this as the configuration is reset
before each test and resetting it after tests will slow down the
integration tests.

[1]:
https://github.com/alphagov/vcloud-edge_gateway/blob/master/spec/integration/edge_gateway/configure_services_spec.rb#L92-L103
[2]:
https://github.com/alphagov/vcloud-edge_gateway/commit/52b1b7eba7e2ddf3b74afd254ee59c0e8f4cb157#diff-29621b5d73d46c66fbeb9a817a3fa252R43