Add #network test; account for placeholder network
The vCloud API returns a weird 'none' vAppNetwork in the output of
Vapp#networks. I've documented this in #72073214, but for the record
it appears that the output of
`vcloud_attributes[:'ovf:NetworkSection'][:'ovf:Network']` always contains
a network called 'none', detailed as a 'placeholder':

````
  :ovf_name => "none",
  :"ovf:Description" => "This is a special place-holder used for disconnected network interfaces."
````

This appears regardless -- if there are actual vAppNetworks defined, it will still appear.

I've added a helper method -- `sanitize_network_output` -- prevent this behaviour from
muddying each test on the number of networks expected.