Add 'bogus networks' case for Vapp#instantiate
If networks are specified to Vapp#instantiate that are not present
in the target organization, the #instantiate call will succeed, simply
skipping the bogus networks.

This is arguably a bug, that we should instead raise an error.

I have logged #70363690 previously to fix this, but it is worth
documenting the behaviour in the tests.