Add missing get_vdc delegator, discovered by test
Splitting this into its own commit, to highlight that I'm
fixing a bug discovered whilst writing an integration test using Vdc#name

When running them initially, I got:

````
NoMethodError:
  undefined method `get_vdc' for #<Vcloud::Fog::ServiceInterface:0x007ff433291920>
````

This was due to a missing FogFacade delegator for get_vdc. This bug must have been
present for a while, indicating that we are not actually using Vdc.name or
Vdc.vcloud_attributes anywhere.