Replace methadone with optparse and tests
This is essentially the same as alphagov/vcloud-edge_gateway#70. The
executable has been rewritten to use plain OptionParser and unit tested.

With the difference that exceptions are caught to prevent the stacktrace
being leaked to the CLI, because there's no validation of "types" until you
hit the API, and the stacktrace you get back through Core/Fog/Excon is
exceptionally long.

Aruba/Cucumber has been removed because we no longer need this to
superficially test the command line interface.