Prefer `@test_data` hash to multiple `let`s
Instead of using `let` to define the variables needed in the
`EdgeGateway` integration tests, use the `@test_data` instance variable.

This should simplify the code somewhat and also avoids having `let`
declarations be specific to the scope of a single `context` block
unnecessarily (which reduces their visibility and thus the readability
of the code).