Rename Rubocop
As of version 0.23.0[1], Rubocop is now RuboCop. This broke our CI
builds.

Fix the namespacing in our Rakefile to reflect this and use pessimistic
versioning to pin Rubocop to the 0.23.x series in our gemspec file.
Since RuboCop's API is not yet stable (as would be the case from version
1.0.0), future changes such as these could break our builds again.

[1]: bbatsov/rubocop@7ee7a8b