Use new TestSetup interface to access test params
As of version 0.1.0 of the vCloud Tools Tester gem, we must use the
`Vcloud::Tools::Tester::TestSetup#test_params` method to retrieve test
parameters.

Also, update the gemspec file to reflect this.

* * *

Note that I've intentionally left the `expected_user_params` argument
for `TestSetup#new` empty; this is new functionality that was added in
vCloud Tools Tester version 0.1.0 and is outside the scope of the
current story.