Allow vcloud-login password to be piped in
The highline gem pretty much does this for us. All we need to do is change
the prompt to make it clear that it's reading data, and not output the
masked password. This will make it easier for us to use this login method in
CI where Jenkins has the password as an environment variable.