Suppress noisy fog "missing credentials" error
This occurs if we call ::Fog.credentials and it has been unable to load any
credentials from a FOG_RC file. If this happens then we can assume that
there isn't a plaintext password set, so continue. A more succinct error
will be produced when we instantiate Fog::Compute::VcloudDirector in
get_token().

This is an integration test rather than a unit test because we're testing
the behaviour of Fog (with some small instrumentation).