Make Vcloud::Fog::Login methods really private
The check_plaintext_pass and get_token methods weren't actually private
before because the `private` method has no effect on module methods.

I'm using `subject.token` as the public entry point instead of
`subject.send(:check_plaintext_pass, ...)` because it matches the normal
usage and also asserts that we don't attempt to get the token if we find a
plaintext password.

Thanks @timmow for the advice!