Replace token length regex with string#size
We don't need to use a fancy regex to check the length of a string. This is
much easier to read. I've broken the first token test out to a separate var
so that we're not doing too much with `expect()`.

NB: String length is our best indicator of a valid token because the
    contents are generated randomly.