Use vcloud-login for integration tests
Writes a temporary FOG_RC containing the API host and username which is used
by both vcloud-login and the integration tests.

This re-instates 5ebf8aa but without the typo that was corrected in a045304.
It was reverted because the token was echoed to STDOUT and logged by
Jenkins. I'll address this in a separate commit.