Deprecation warning for vcloud_director_password
Raise a warning (to STDERR) when vcloud/core is loaded and Fog credentials
has :vcloud_director_password (which will have come from a plaintext FOG_RC
file).

I'm putting the call to check_credentials in the top-scope rather than the
initialisers of Vcloud::Fog::ModelInterface & Vcloud::Fog::ServiceInterface
so that we don't forget to add it to additional Fog interfaces should they
exist in the future.

This is in anticipation of actively preventing the use of passwords in
plaintext files for future versions of vcloud-tools now that we have the
vcloud-login utility.