Reinstate Login#token integration tests
This reverts commit ad15254 which moved these to a file that wouldn't be
picked up by rspec's glob. They have been modified so as not to conflict
with what we discovered in 09ef65f and e766457.

We're able to run these in CI because we still have access to the original
password injected by Jenkins which is used by `vcloud-login` in
`jenkins_tests.sh`. When not run in CI and the environment variable isn't
set, these tests will be marked as pending.