Ban plaintext passwords
In version 0.7.0 we deprecated the use of plaintext passwords in FOG_RC
in a conscious move to promote that as bad practise. Interaction with
vcloud-tools should now be done by supplying a token, which can be
gained using `vcloud-login` (which will accept a password on STDIN).

This involved:

 - Moving check_plaintext_pass into Vcloud::Core::Fog
 - Calling check_plaintext_pass from check_credentials
 - replacing call to check_plaintext_pass from Vcloud::Core::Fog::Login
   with call to Vcloud::Core::Fog.check_credentials
 - Rescuing call to Vcloud::Core::Fog.check_credentials in
   `vcloud/core/fog.rb` to provide a nice error message to the user

Happy to be guided on better ways to do this - it's my first solo foray
into vcloud-tools.

Rubocop and Specs pass. Not much changed in the test apart from the
expected error message, which I have lengthened slighly to describe the
reason behind the error as well as what it means.