Add basic support for Independent Disks
Independent Disks are their own vCloud Entity, unrelated to the 'local' disks
that can currently be added to VMs via the Vm.extra_disks method. They are conceptually
similar to AWS EBS volumes.

Instead, independent disks are created as their own entity, and *attached*
to a VM. This can be done when the VM is stopped or running.

When the VM is destroyed, any attached Independent Disks will be detached,
but not destroyed. They can subsequently be re-attached to another (usually recreated) Vm.

This commit adds basic support for referring to existing Independent Disks. Future commits
will add the necessary create and attach operations.