Allow specification of size param in GB/MB/MiB/GiB
It is very impractical to expect folk to specify a disk size
in bytes, when it is far more commonplace to specify them in GB/MB/etc.

Disks are usually treated as 1MB = 10^6 bytes rather than 2^20, so use
this convention but also allow for 'MiB' and 'GiB' to correctly specify
2^20 and 2^30 respectively.

Continue to support bytes being supplied as a string or an Integer.

Raise an ArgumentError if we cannot parse the input string.