Add IndependentDisk#attached_vms; disk integ tests
IndependentDisk#attached_vms allows us to know where a disk has been attached
to, so we can check that it has been attached correctly.

This method allows us to reliably test the attachment of disks to VMs, so an
integration test has been added.

NB: It is possible to detect Independent Disks attached to a Vm via the
Vm#vcloud_attributes call, by looking for a RASD OVF disk with a :vcloud_disk key,
but this is quite an error-prone and ugly process.