Add Vapp.get_by_child_vm_id constructor
At present, to create a Vm object you must also supply the parent
vApp object. This is technically unneccessary, since the vm retrieved
from the API knows this calready.

In the interests of not changing the interface of the Vm.new method
though, we must provide a way of discovering the parent vApp in order
to create a Vm object from a vm_id reference.

This work was prompted by the need to create a bunch of Vm object
from the output of the 'get_vms_disk_attached_to' Fog method, which
returns a list of Vm ids that a given independent_disk_id is attached
to. I could not return a list of Vm objects without a way to find the
parent Vapp of each.