Improve multiple-disks error message
IndependentDisks#get_by_name_and_vdc_name cannot handle the case where
multiple disks have the same name, in the same vdc. We try to prevent
this, but we cannot prevent it in the API. If the situation arises, the
IndependentDisk API can still be used, but the disks must be referenced
using the .new(id) constructor directly.