Fix Vm integ test via #detach_independent_disks
PR https://github.com/gds-operations/vcloud-core/pull/130 broke the integration
tests due to an oversight with Storage Profiles.

It appears that when Independent Disks are attached, it is not possible to change
the storage profile of the attached VM. Errors occur via both the API and UI.

As a result, I have created the Core::Vm#detach_independent_disks call, which works
in the same manner as #attach_independent_disks -- takes a list of IndependentDisk objects
to detach.

The vm_spec.rb integration test has been updated to document that errors will occur if
independent disks are attached, and that they should be detached first.