Manual pagination for page/pageSize in QueryRunner
Disable automatic pagination when the `page` and/or `pageSize` options are
passed to `QueryRunner#run`. The previous use of `get_all_results` meant
that:

- `page` was stripped out of the options hash and we always began at page 1
- `pageSize` was honoured, but all pages would be fetched, resulting in many
  smaller and inefficient queries

Allowing the use of these options will allow consumers to make much more
efficient queries. The first use case for this is [#70584000] and
gds-operations/vcloud-edge_gateway#104 where I want to fetch the single most
recent task by using `page: 1, pageSize: 1, sortDesc: startDate`.

It's not feasible to integration test the `page` option without `pageSize`
because we'd need a result set bigger than the default of 25.