Mark missing credentials test as pending
This test is currently failing on `master` due to changes upstream in the
fog-core gem.

Specifically, fog/fog-core@08df305 changed the
`::Fog::Errors::LoadError` class but omitted to update a `rescue` line
that refers to that class.

I've raised a pull request against fog-core to fix that issue:
https://github.com/fog/fog-core/pull/97

In the meantime, to fix `master` in a timely manner, mark this test as
pending while we wait for a new release of `fog-core` including the fix
in that PR.