Re-work test namespacing; removes extraneous k_END
Namespace calls to Vcloud::Core in tests. Currently, tests have been
namespaced as thus:

Module Vcloud
  Module Core
    Describe blah

Here, we namespace them as:

Describe Vcloud::Core::blah

We do this to ensure consistency across namespacing calls as it is currently
possible to call an object without properly namespacing it. This means
they're likely to get missed when refactored, or similar.

Similarly, refactoring them makes them easier to read, which is good.