Explicitly set fog-core version
We wouldn't normally specificy a `fog-core` version because it's already
pulled in by `fog` [1] (and thus not our concern).

But in this case, we set it explicitly so that we can ensure we use
version 1.27.4 and above, which includes this bugfix:

https://github.com/fog/fog-core/commit/58556e4b1e65d2ac8dd177c531dd8a23929d39e3

Without that bugfix, our integration tests (`bundle exec rake integration`) were failing with this error:

    NoMethodError:
          undefined method `redisplay_progressbar' for Fog::Formatador:Class

Note that we change the version pinning slightly to use pessimistic
versioning to prevent us from inadvertently using version 2 of
`fog-core`, while ensuring that we use at least version 1.27.4.

See also the original bug report for vCloud Tools:
https://groups.google.com/a/digital.cabinet-office.gov.uk/d/msg/vcloud-tools/GF-YrOVhUUU/bSQVpWyuaQkJ

[1]: https://github.com/fog/fog/blob/16652f124f22f0d7f4addd44f4bb50dde152bb60/fog.gemspec#L49