Don't test for length of vCloud login token
Prior to 36ba0cc1, we asserted that the length of a vCloud Director
login token must be 44 characters.

In 36ba0cc1 we changed that length to 32 characters, I suspect because
a later version of vCloud Director returns token of that length.

Our integration tests are current passing in one provider (Skyscape) but
failing in another (Carrenza), presumably due to a discrepancy in the
versions of vCloud Director they use. Carrenza is returning a token 44
characters long.

Since the exact format of the token is undocumented [1], don't check the
length and only assert that we received a non-empty string. Otherwise,
we risk having to change this value each time VMware change its format.

http://pubs.vmware.com/vcd-56/topic/com.vmware.ICbase/PDF/vcd_56_api_guide.pdf