Remove support for Ruby 1.9.3
Official support for Ruby 1.9.3 ended on February 23, 2015:
https://www.ruby-lang.org/en/news/2014/01/10/ruby-1-9-3-will-end-on-2015/

Version 3.0.1 of net-ssh, a dependency of fog-core, no longer supports
Ruby 1.9.3:
https://github.com/net-ssh/net-ssh/blob/edd7335d42630aec1ad5b18542e863608bb4ed39/CHANGES.txt#L27

There's a fix for this in fog-core, which is not yet released:
https://github.com/fog/fog-core/pull/161

Rather than wait for the above fix to be released, remove support for
Ruby 1.9.3 on the basis that it's no longer supported.

Tell Jenkins to explicitly run our integrations on the oldest supported
version (consistent with our `.travis.yml`), Ruby 2.0.0. Use the
`RBENV_VERSION` environment variable rather than creating a
`.ruby-version` to avoid users new to Ruby/vCloud Tools from thinking
that 2.0.0 is the only version we support.