Upgrade fog from at least 1.34 to at least 1.36
There's a bug in fog 1.34 and 1.35 where `net/ssh` wasn't removed
properly [1] which is causing problems. Bumping the requirement to
at least 1.36 fixes the issue.

However bumping the version of fog causes our tests to fail with the
error:

```
     NameError:
       uninitialized constant Vcloud::Core::Logger
     # ./lib/vcloud/core.rb:31:in `logger'
     # ./spec/vcloud/core/config_loader_spec.rb:54:in `block (3 levels) in <top (required)>'
```

I'm not sure why this upgrade has caused that to break, but explicitly requiring
the Logger class in that module fixes it.

[1]: https://github.com/fog/fog/issues/3765